move date format timezone detection to base adapter
